Updated in conjuction with Zuxelus' source. New stuff added, works much better as well.

+++ src/main/java/shedar/mods/ic2/nuclearcontrol/api/ICardSettingsWrapper.java
@@ -9,7 +9,7 @@ package shedar.mods.ic2.nuclearcontrol.api;
  */
 public interface ICardSettingsWrapper{
 	void setInt(String name, Integer value);
-	void setLong(String name, Long value);
+	void setDouble(String name, double value);
 	void setString(String name, String value);
 	void setBoolean(String name, Boolean value);
