+++ src/main/java/shedar/mods/ic2/nuclearcontrol/crossmod/data/EnergyStorageData.java
@@ -6,8 +6,8 @@ public class EnergyStorageData
     public static final int TARGET_TYPE_IC2 = 0;
     public static final int TARGET_TYPE_BC = 1;
     
-    public long stored;
-    public long capacity;
+    public double stored;
+    public double capacity;
     public String units;
     public int type;
